+++ core/clothing-store/src/main/java/pro/redsoft/clothingstore/repository/CategoryRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
  * @author Alexander Novik
  *         Date: 30.11.13
  */
-public interface CategoryRepository extends CrudRepository<Category,Integer> {
+public interface CategoryRepository extends CrudRepository<Category,Long> {
 
     Category findByTitle(String title);
