+++ core/clothing-store/src/main/java/pro/redsoft/clothingstore/repository/ContactRepository.java
@@ -7,5 +7,5 @@ import pro.redsoft.clothingstore.domain.Contact;
  * @author Alexander Novik
  *         Date: 28.11.13
  */
-public interface ContactRepository extends CrudRepository<Contact,Integer> {
+public interface ContactRepository extends CrudRepository<Contact,Long> {
 }
