+++ core/clothing-store/src/main/java/pro/redsoft/clothingstore/repository/OrderRepository.java
@@ -7,5 +7,5 @@ import pro.redsoft.clothingstore.domain.order.Orders;
  * @author Alexander Novik
  *         Date: 29.11.13
  */
-public interface OrderRepository extends CrudRepository<Orders,Integer> {
+public interface OrderRepository extends CrudRepository<Orders,Long> {
 }
